requirements.txt
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==2.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.9
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydot==3.0.4
PyJWT==2.10.1
pyparsing==3.2.1
pytest==7.4.3
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
